Refactoring
- Deleted extra unused pages/ pictures from past project.
- Renamed some variables for better clarity
- Rewrote login/ signup pages code for better readability
- Removed unnecessary code, updated widgets and removed the weird measurement unit of fem and ffem.

Note: Both the unnatural number of pages and images, and the measurement unit of fem and ffem were the result of a "Figma to design" converter we tried to use to ease some of the load on our small team.

The result was abysmal and we ended up rewriting most pages, though we didn't get rid of the extra ones since my partner thought we could make use of them down the line.

For the weird units (fem ffem), I initially thought they were a standard measurement in flutter development for responsive design, I couldn't understand nor find out what they did, so I didn't touch them until now.

ChatGPT finally helped solve the mystery; the units were a hacky way to fit the static Figma design into the screen of a device, using the difference of width between the design and that of the device as a guide to position the elements on screen based on the new width/height.